feat : Cars 클래스에서 Record 클래스로 분리
- Cars 클래스에서 Record 클래스로 분리
